+++ designer/client/src/store/configureStore.ts
@@ -21,51 +21,56 @@ const actionsBlacklist: Action["type"][] = [
     "DISPLAY_LIVE_DATA",
 ];
 
-const listenerMiddleware = createListenerMiddleware<RootState, ThunkDispatch<RootState, unknown, Action>>();
+export const setupStore = () => {
+    const listenerMiddleware = createListenerMiddleware<RootState, ThunkDispatch<RootState, unknown, Action>>();
 
-export const store = configureStore({
-    reducer: rootReducer,
-    middleware: (getDefaultMiddleware) =>
-        getDefaultMiddleware({
-            serializableCheck: false, // we still have non fixed antipatterns
-            thunk: false, // need to disable and provide own, typed thunk
-        })
-            .prepend(listenerMiddleware.middleware, thunk as ThunkMiddleware<RootState, Action>)
-            .concat(
-                createStateSyncMiddleware({
-                    whitelist: [
-                        "TOGGLE_SETTINGS",
-                        "SET_SETTINGS",
-                        "REGISTER_TOOLBARS",
-                        "RESET_TOOLBARS",
-                        "MOVE_TOOLBAR",
-                        "TOGGLE_TOOLBAR",
-                        "TOGGLE_ALL_TOOLBARS",
-                        "TOGGLE_PANEL",
-                        "TOGGLE_COMPONENT_GROUP_TOOLBOX",
-                    ],
-                }),
-                nodeValidationMiddleware([
-                    "NODE_ADDED",
-                    "DELETE_NODES",
-                    "NODES_CONNECTED",
-                    "NODES_DISCONNECTED",
-                    "NODES_WITH_EDGES_ADDED",
-                    "STICKY_NOTE_UPDATED",
-                ]),
-            ),
-    devTools: {
-        actionsDenylist: ["RNS_SHOW_NOTIFICATION", "RNS_HIDE_NOTIFICATION", ...actionsBlacklist],
-    },
-});
-
-export const persistor = persistStore(store as Store);
-initMessageListener(store);
-
-if (module.hot) {
-    module.hot.accept("../reducers", () => {
-        // eslint-disable-next-line @typescript-eslint/no-var-requires
-        const nextReducer = require("../reducers").reducer;
-        store.replaceReducer(nextReducer);
+    const store = configureStore({
+        reducer: rootReducer,
+        middleware: (getDefaultMiddleware) =>
+            getDefaultMiddleware({
+                immutableCheck: false,
+                serializableCheck: false, // we still have non fixed antipatterns
+                thunk: false, // need to disable and provide own, typed thunk
+            })
+                .prepend(listenerMiddleware.middleware, thunk as ThunkMiddleware<RootState, Action>)
+                .concat(
+                    createStateSyncMiddleware({
+                        whitelist: [
+                            "TOGGLE_SETTINGS",
+                            "SET_SETTINGS",
+                            "REGISTER_TOOLBARS",
+                            "RESET_TOOLBARS",
+                            "MOVE_TOOLBAR",
+                            "TOGGLE_TOOLBAR",
+                            "TOGGLE_ALL_TOOLBARS",
+                            "TOGGLE_PANEL",
+                            "TOGGLE_COMPONENT_GROUP_TOOLBOX",
+                        ],
+                    }),
+                    nodeValidationMiddleware([
+                        "NODE_ADDED",
+                        "DELETE_NODES",
+                        "NODES_CONNECTED",
+                        "NODES_DISCONNECTED",
+                        "NODES_WITH_EDGES_ADDED",
+                        "STICKY_NOTE_UPDATED",
+                    ]),
+                ),
+        devTools: {
+            actionsDenylist: ["RNS_SHOW_NOTIFICATION", "RNS_HIDE_NOTIFICATION", ...actionsBlacklist],
+        },
     });
-}
+
+    const persistor = persistStore(store as Store);
+    initMessageListener(store);
+
+    if (module.hot) {
+        module.hot.accept("../reducers", () => {
+            // eslint-disable-next-line @typescript-eslint/no-var-requires
+            const nextReducer = require("../reducers").reducer;
+            store.replaceReducer(nextReducer);
+        });
+    }
+
+    return { store, persistor };
+};
