--- designer/client/test/reducer-test.js
@@ -1,217 +1,0 @@
-import { nodeAdded, nodesWithEdgesAdded } from "../src/actions/nk";
-import NodeUtils from "../src/components/graph/NodeUtils";
-import rootReducer from "../src/reducers/index";
-
-const baseProcessState = {
-    name: "DEFGH",
-    isLatestVersion: true,
-    processCategory: "Category1",
-    modificationDate: "2017-02-14T11:16:56.686",
-    createdAt: "2017-02-14T11:16:56.686",
-    createdBy: "admin",
-    tags: [],
-    currentlyDeployedAt: ["test"],
-    scenarioGraph: {
-        properties: {
-            parallelism: 3,
-            additionalFields: {
-                groups: [
-                    {
-                        id: "acdc",
-                        nodes: ["kafka-transaction", "filter"],
-                    },
-                ],
-            },
-        },
-        nodes: [
-            {
-                type: "Source",
-                id: "kafka-transaction",
-                ref: {
-                    typ: "kafka-transaction",
-                    parameters: [],
-                },
-                additionalFields: {
-                    description: "asdfasdfłóóódźźźasdfsdfasdfasdfasdfasdf",
-                },
-            },
-            {
-                type: "Filter",
-                id: "filter",
-                expression: {
-                    language: "spel",
-                    expression: "4 / (#input.length -5) >= 0",
-                },
-            },
-            {
-                type: "Enricher",
-                id: "paramService",
-                service: {
-                    id: "paramService",
-                    parameters: [
-                        {
-                            name: "param",
-                            expression: {
-                                language: "spel",
-                                expression: "'3434'",
-                            },
-                        },
-                    ],
-                },
-                output: "output",
-            },
-            {
-                type: "Sink",
-                id: "sendSms",
-                ref: {
-                    typ: "sendSms",
-                    parameters: [],
-                },
-            },
-        ],
-        edges: [
-            {
-                from: "kafka-transaction",
-                to: "filter",
-            },
-            {
-                from: "filter",
-                to: "paramService",
-                edgeType: {
-                    type: "FilterTrue",
-                },
-            },
-            {
-                from: "paramService",
-                to: "sendSms",
-            },
-        ],
-        stickyNotes: [],
-    },
-    validationResult: {
-        errors: {
-            invalidNodes: {},
-            processPropertiesErrors: [],
-            globalErrors: [],
-        },
-    },
-    state: {},
-};
-
-const baseState = rootReducer(
-    {},
-    {
-        type: "@@INIT",
-    },
-);
-
-const baseStateWithProcess = rootReducer(baseState, {
-    type: "DISPLAY_PROCESS",
-    scenario: baseProcessState,
-});
-
-const reduceAll = (actions) => {
-    let currentState = baseStateWithProcess;
-    const getState = () => currentState;
-
-    const dispatch = (action) => {
-        if (typeof action === "function") {
-            action(dispatch, getState);
-        } else {
-            currentState = rootReducer(currentState, action);
-        }
-    };
-
-    actions.forEach((action) => dispatch(action));
-
-    return currentState;
-};
-
-const testNode = {
-    type: "Enricher",
-    id: "Enricher ID",
-    service: {
-        id: "paramService",
-        parameters: [
-            {
-                name: "param",
-                expression: {
-                    language: "spel",
-                    expression: "'3434'",
-                },
-            },
-        ],
-    },
-    output: "output",
-};
-
-const testPosition = { x: 10, y: 20 };
-const getGraph = (state) => state.graphReducer.present;
-
-describe("Reducer", () => {
-    it("should display process", () => {
-        expect(getGraph(baseStateWithProcess).scenario.name).toEqual(baseProcessState.name);
-    });
-});
-
-describe("Nodes added", () => {
-    it("should add single node", () => {
-        const result = reduceAll([nodeAdded(testNode, testPosition)]);
-
-        expect(NodeUtils.getNodeById(testNode.id, getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(getGraph(result).layout.find((n) => n.id === testNode.id).position).toEqual(testPosition);
-    });
-
-    it("should add single node with unique id", () => {
-        const result = reduceAll([nodeAdded({ ...testNode, id: "kafka-transaction" }, testPosition)]);
-
-        expect(NodeUtils.getNodeById("kafka-transaction 1", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(getGraph(result).layout.find((n) => n.id === "kafka-transaction 1").position).toEqual(testPosition);
-    });
-
-    it("should add multiple nodes with unique id", () => {
-        const action = nodesWithEdgesAdded(
-            [
-                {
-                    node: { ...testNode, id: "kafka-transaction" },
-                    position: { x: 10, y: 20 },
-                },
-                {
-                    node: { ...testNode, id: "filter" },
-                    position: { x: 10, y: 20 },
-                },
-            ],
-            [],
-        );
-        const result = reduceAll([action, action]);
-
-        expect(NodeUtils.getNodeById("kafka-transaction (copy 1)", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(NodeUtils.getNodeById("kafka-transaction (copy 2)", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(NodeUtils.getNodeById("filter (copy 2)", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-    });
-
-    it("should add nodes with edges", () => {
-        const result = reduceAll([
-            nodesWithEdgesAdded(
-                [
-                    {
-                        node: { ...testNode, id: "newNode" },
-                        position: testPosition,
-                    },
-                    {
-                        node: { ...testNode, id: "kafka-transaction" },
-                        position: testPosition,
-                    },
-                ],
-                [{ from: "newNode", to: "kafka-transaction" }],
-            ),
-        ]);
-
-        expect(NodeUtils.getNodeById("newNode", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(NodeUtils.getNodeById("kafka-transaction (copy 1)", getGraph(result).scenario.scenarioGraph)).toMatchSnapshot();
-        expect(NodeUtils.getEdgeById("newNode-kafka-transaction (copy 1)", getGraph(result).scenario.scenarioGraph)).toEqual({
-            from: "newNode",
-            to: "kafka-transaction (copy 1)",
-        });
-    });
-});
